formatting and adjusting the bit width of wire perf_rdata

+++ app/util.h
@@ -4,5 +4,5 @@
 void pg_exit();
 void pg_printc(char c);
 void pg_printd(long long x);
-void pg_printh(int x);
+void pg_printh(unsigned long x);
 void pg_prints(const char *str);
